Вынес всю логику редактирования фото в отд. модуль
