+++ js/main.js
@@ -1,3 +1,4 @@
 import './rendering-mini.js';
 import './rendering-full.js';
 import './upload.js';
+import './redactor.js';
